Feature/20 support experimental endpoints delete export

+++ averbis/core/_rest_client.py
@@ -24,6 +24,7 @@ import logging
 from concurrent.futures.thread import ThreadPoolExecutor
 from functools import wraps
 from io import BytesIO
+from json import JSONDecodeError
 
 from time import sleep, time
 from typing import List, Union, IO, Iterable, Dict, TextIO, Iterator, Optional
@@ -204,15 +205,21 @@ class Pipeline:
         # noinspection PyProtectedMember
         return self.project.client._get_pipeline_info(self.project.name, self.name)
 
-    def delete(self) -> None:
+    @experimental_api
+    def delete(self) -> dict:
         """
-        Delete the pipeline from the server.
+        HIGHLY EXPERIMENTAL API - may soon change or disappear. Deletes an existing pipeline from the server.
 
         :return: The raw payload of the server response. Future versions of this library may return a better-suited
                  representation.
         """
+        if self.project.client.spec_version.startswith("5."):
+            raise OperationNotSupported(
+                "Deleting pipelines is not supported by the REST API in platform version 5.x, but only from 6.x onwards."
+            )
+
         # noinspection PyProtectedMember
-        self.project.client._delete_pipeline(self.project.name, self.name)
+        return self.project.client._delete_pipeline(self.project.name, self.name)
 
     def is_started(self) -> bool:
         """
@@ -633,6 +640,25 @@ class Project:
         # noinspection PyProtectedMember
         return self.client._select(self.name, query, **kwargs)
 
+    @experimental_api
+    def export_text_analysis(
+        self, document_sources: str, process: str, annotation_types: str = None
+    ) -> dict:
+        """
+        HIGHLY EXPERIMENTAL API - may soon change or disappear. Exports a given text analysis process as a json.
+
+        :return: The raw payload of the server response. Future versions of this library may return a better-suited
+         representation.
+        """
+        if self.client.spec_version.startswith("5."):
+            raise OperationNotSupported(
+                "Text analysis export is not supported for platform version 5.x, it is only supported from 6.x onwards."
+            )
+        # noinspection PyProtectedMember
+        return self.client._export_text_analysis(
+            self.name, document_sources, process, annotation_types
+        )
+
 
 class Client:
     def __init__(
@@ -641,6 +667,8 @@ class Client:
         api_token: str = None,
         verify_ssl: Union[str, bool] = True,
         settings: Union[str, Path, dict] = None,
+        username: str = None,
+        password: str = None,
     ):
         self.__logger = logging.getLogger(self.__class__.__module__ + "." + self.__class__.__name__)
         self._api_token = api_token
@@ -658,7 +686,22 @@ class Client:
                 self._apply_profile("*")
             self._apply_profile(url_or_id)
 
-        self.build_info = self.get_build_info()
+        if self._api_token is None:
+            if username is not None and password is not None:
+                self.regenerate_api_token(username, password)
+            else:
+                raise Exception(
+                    "An API Token is required for initializing the Client.\n"
+                    + "You can either pass it directly with: Client(url,api_token=your_token) or you can\n"
+                    + "generate a new API token with: Client(url, username='your_user_name', password='your_password')."
+                )
+
+        try:
+            self.build_info = self.get_build_info()
+        except JSONDecodeError:
+            raise ValueError(
+                "The Client could not get information about the platform. This is likely because your API Token has changed."
+            )
         self.spec_version = self.build_info["specVersion"]
 
     def _exists_profile(self, profile: str):
@@ -1052,11 +1095,15 @@ class Client:
         )
         return response["payload"]
 
-    def _delete_pipeline(self, project: str, pipeline: str) -> None:
+    @experimental_api
+    def _delete_pipeline(self, project: str, pipeline: str) -> dict:
         """
         Use Pipeline.delete() instead.
         """
-        raise OperationNotSupported("Deleting pipelines is not supported by the REST API yet")
+        response = self.__request(
+            "delete", f"/experimental/textanalysis/projects/{project}/pipelines/{pipeline}"
+        )
+        return response["payload"]
 
     def _start_pipeline(self, project: str, pipeline: str) -> dict:
         response = self.__request(
@@ -1163,6 +1210,21 @@ class Client:
             f"/v1/search/projects/{project}/select",
             params={"q": q, **kwargs},
             headers={HEADER_CONTENT_TYPE: MEDIA_TYPE_TEXT_PLAIN_UTF8},
+        )
+        return response["payload"]
+
+    @experimental_api
+    def _export_text_analysis(
+        self, project: str, document_sources: str, process: str, annotation_types: str = None
+    ):
+        """
+        Use Project.export_text_analysis() instead.
+        """
+        response = self.__request(
+            "get",
+            f"/experimental/textanalysis/projects/{project}/documentSources/{document_sources}/processes/{process}/export",
+            params={"annotationTypes": annotation_types},
+            headers={HEADER_ACCEPT: MEDIA_TYPE_APPLICATION_JSON},
         )
         return response["payload"]
